requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
qrcode==7.4.2
pypng==0.20220715.0
sqlparse==0.5.3
typing_extensions==4.14.0
phonenumbers==9.0.12

# Deployment & DB
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Utilities
requests==2.32.3
python-dotenv==1.0.1
